requirements.txt
discord.py==2.3.2
gspread==5.12.4
oauth2client==4.1.3
